refactor: store sim output in Vec, fn to make Kanata from str
Add `new_from_str` for `Kanata` and `Cfg`. The goal is to move towards
being able to run a Kanata state machine on the web, which does not have
access to files.

Store simulated output strings in a `Vec` instead of printing to stdout
immediately. The goal is to give control to the consuming application
for where to output the text, e.g. to a web text box instead of stdout.