feat!: check all held layers for transparent action
Add a breaking-but-revertable change to resolve transparent actions more
intuitively. Kanata used to only check the base layer for transparent
actions. It also did not resolve transparent actions nested within more
complex actions like `tap-hold`.

This commit makes all actions resolve transparency correctly; though a
transparent chord action that involves multiple keys is not really
sensible. That may be worth blocking in the parser in the future.

The breaking change is that all while-held layers are checked for
transparent action resolution instead of only the base layer. This check
is done in most-recent-first order. The breaking behaviour can be
reverted using a `defcfg` item.