fix!: restrict TCP server to localhost by default
I read that using `127.0.0.1` as opposed to `0.0.0.0` (or any other IP
address) prevents TCP connections to the server from outside the
machine, using `127.0.0.1` (the loopback interface IP address) limits
the connections to connections from the local machine, even if the
machine's firewall was misconfigured or for some reason is not working.

This change improves security by default. The default is a breaking
change but users who have been connecting to kanata over the network and
not just via localhost can opt out of the default. Presumably few if any
users were connecting to kanata outside of local processes.