feat: chords v2 experimental
Chords v2 implements global chords with an option to disable chords on
specific layers. In the implementation, inputs first go into a separate
queue for chords v2 processing. This is in contrast to the existing
chords implementation, which uses the same Layout queue as the rest of
the keyberon actions.

An advantage of this system is that this enables using the same chord
definition on multiple layers without manually changing/defining the key
actions of individual keypresses. To contrast the original
implementation, actions for an individual key press must also be defined
within the chord group.

A notable addition that is adjacent to the chords implementation is the
use of simulated tests which work similarly to the existing keyberon
tests, but which I find easier to read and write.