fix!(transparent): forbid nested trans
This commit forbids nested transparent actions on the first layer when
`delegate-to-first-layer` is enabled.

This change is done because, with the way the current code works, this
configuration causes an infinite recursion leading to stack overflow.
The behaviour was introduced in commit 089853a which improves the
runtime memory consumption of kanata by having only one keyberon layer
for each configured layer, instead of having two variants (base +
while-held). One of the changes necessary to remove the two layer
variants was to modify the `defsrc_layer` with the content of the first
defined layer, and it is this change that causes the infinite recursion.
It seems undesirable to revert the improvement to enable this niche, and
in my opinion, nonsensical use case. This same use case also already
behaves incorrectly v1.5.0, so there is no loss in functionality from a
non-prerelease version.

Some considerations was made to fix the behaviour by adding to the
transparency resolution code, rather than reverting commit 089853a. For
example, introduce a second, deeper `src_layer` to the keyberon `Layout`
which would be the original `defsrc` layer, with no modification from
`delegate-to-first-layer`. However, this code and memory bloat does not
seem worthwhile to enable the niche+nonsensical use case.