make manifest build optional
This is done because this change adds 30 new crates which need to be
downloaded and compiled. It seems wasteful to do this unless explicitly
opted into, which can be done for those that desire it and for official
builds.

Additionally, the number of `cfg` declarations is minimized in build.rs
by declaring an in-file module.