feat: allow `t!` in place of `template-expand` (#958)
Use `t!` as a short form of `template-expand`.

The alternative of allowing the template name in the first atom was
considered. However, such an implementation would be more difficult and
complex. Allowing a shorter prefix atom is a simple and quick
implementation.

If one wanted to do the work in the future to allow even shorter
template expansion by adding syntax to remove the prefix atom
altogether, the added code should not conflict with that work. For now
though, the simpler implementation will be added for easier maintenance
and hopefully less risk of introducing bugs.