ci: adjust how jobs run (#974)
Split up pipeline jobs by operating system.

The macOS pipeline has been inconsistently failing for some reason, with
an incomprehensible SIGABRT of all things. The main goal of this PR is a
shot-in-the-dark that reducing the stuff that runs in the macOS pipeline
will help with the consistency.

It's also a generally good improvement anyway because Linux and macOS
don't use a bunch of the feature flags that Windows uses.

Some other miscellaneous test+lock changes made it in too, not worth
undoing that.