feat: add special noop keys (#961)
Add no-op keys that are never output as OS events but can participate in
sequences and fork/switch logic.

The keys are mapped to a hopefully low-risk set of already-mapped
codes (for Linux at least). The advantage of using already-mapped codes
is that they won't change purpose in the future. The disadvantage is of
course that there are already known uses of these keys. The known
purpose is presumed to be low risk for negatively impacting users in
practice, and a workaround is to not map keys from `defsrc`.