feat: unblock and forward buttons properly (#972)
Closes #968. For now there is no way to block mouse buttons that are
being forwarded to kanata. If someone requests it or opens a PR, the
feature may be included, but for now I don't see a reason why it's
necessary. It seems much more sensible for someone to want to block
their unused keyboard buttons to force the learning of new typing habits
compared to blocking mouse buttons for a similar reason.