fix(repeat): check event code itself for repeat output (#986)
This commit fixes key repeat in Linux TTY and Windows when a key on the
base layer is transparent, OR process-unmapped-keys is enabled and the
key is unmapped.

This functionality was broken in commit 089853a which changed how layers
worked by removing the duplicated base-version and while-held-version of
layers, for the benefit of reduced memory consumption.