feat: definitions and references lsp hint (1016)
Add goto definition support for LSP consumers. Adding support for it
directly in LSP server could *maybe* be doable, but unreliable (e.g.
templates are complicating things) and so a lot of parser logic would
need to be reimplemented for all edge cases, so I think this is better
added directly to kanata-parser.