fix(win-llhook): restore MapVirtualKeyA for default kanata.exe
The commit fab0d89 intended to fix the functionality of keypad enter.
It worked for the US layout but broke other OS layouts. It turns out
MapVirtualKeyA **was** needed to translate to the appropriate scancode
output. The fix for keypad enter is done by attempting `osc_to_u16` if
MapVirtualKeyA fails, which it should for a nonexistent VK that kanata
uses for keypad enter.